Add Win32 layer, package manager, and scripting engine (Phases 34-36)
Phase 34 - Win32 compatibility: PE32 loader plus kernel32/user32/gdi32/advapi32
shim (47 functions) and in-memory registry emulation (win32.c/h, registry.c/h;
win32info/regedit/runexe shell commands).

Phase 35 - Package manager: binary .npk archive format with CRC32 integrity
(header + per-file), a bundled repository, recursive dependency resolution with
cycle detection and minimum-version checks, and an install/remove/update engine
that serializes -> verifies -> extracts into the VFS with rollback (pkg.c/h;
npkg list/search/info/install/remove/installed/update).

Phase 36 - Scripting engine: NexusScript interpreter (lexer + recursive-descent
evaluator over a token cursor) with int/string values, variables, arithmetic
and logical expressions, if/elif/else and while, len/str/abs builtins, and
run <cmd> for OS automation, guarded by a step budget so a runaway script
cannot hang the kernel (script.c/h; script <file.ns>, shell_exec_line hook).

Wires all three into kernel_main init, the shell dispatch/help, and the desktop
terminal; README/roadmap updated through Phase 36 (36/51, 71%). Also includes
in-progress kernel/desktop working-tree changes (boot stack layout, icons,
start menu, vga, wallpaper, window).

+++ NexusOS/kernel/vga.c
@@ -31,7 +31,7 @@ static bool vesa_mode = false;
  * Make a VGA entry: char + color packed into 16 bits
  * -------------------------------------------------------------------------- */
 static uint16_t vga_entry(char c, uint8_t color) {
-    return (uint16_t)c | ((uint16_t)color << 8);
+    return (uint16_t)(uint8_t)c | ((uint16_t)color << 8);
 }
 
 /* --------------------------------------------------------------------------
